requirements.txt
pymysql==1.1.0
pandas==2.2.2
dolphindb>=3.0
loguru==0.7.2
pyyaml==6.0
keyring>=25.0
